+++ constants/Jurassic World Evolution/mimes.py
@@ -1,4 +1,4 @@
-from ovl_util.mimes import Mime
+from constants import Mime
 
 mimes = {
 	'.achievements': Mime(name='Casino:Achievements:achievements', hash=1130939801, version=1, triplets=[], pool=4, set_pool=0),
