+++ gui/setup/bnk_gui_setup.py
