+++ gui/setup/fgm_editor_setup.py
