+++ gui/tooltips/fgm.txt
