+++ version.txt
@@ -1,1 +1,1 @@
-843692cc - Sun Jul 30 19:47:26 2023 -0400
+595c0e33 - Sun Jul 30 19:58:12 2023 -0400
